RavenDB-20565 - make sure that we aren't running a query while we clean the lucene cache

+++ src/Raven.Server/Indexing/LuceneCleaner.cs
@@ -1,10 +1,15 @@
-﻿using Lucene.Net.Search;
+﻿using System;
+using System.Threading;
+using Lucene.Net.Search;
+using Raven.Client.Util;
 using Sparrow.LowMemory;
 
 namespace Raven.Server.Indexing;
 
 public class LuceneCleaner : ILowMemoryHandler
 {
+    private readonly ReaderWriterLockSlim _runningQueryLock = new();
+
     public LuceneCleaner()
     {
         LowMemoryNotification.Instance.RegisterLowMemoryHandler(this);
@@ -12,10 +17,31 @@ public class LuceneCleaner : ILowMemoryHandler
 
     public void LowMemory(LowMemorySeverity lowMemorySeverity)
     {
-        FieldCache_Fields.DEFAULT.PurgeAllCaches();
+        if (_runningQueryLock.TryEnterWriteLock(10) == false)
+            return;
+
+        try
+        {
+            // PurgeAllCaches is replacing the cache with a new one (without actually releasing any memory).
+            // When the GC will run, the finalizer of the Segments will be executed and release the unmanaged memory.
+            // HOWEVER, this will happen when the managed memory is high enough to trigger a GC cycle.
+            // Which is too late since we are already in a low memory state
+            FieldCache_Fields.DEFAULT.PurgeAllCaches();
+        }
+        finally
+        {
+            _runningQueryLock.ExitWriteLock();
+        }
     }
 
     public void LowMemoryOver()
     {
+    }
+
+    public IDisposable EnterRunningQueryReadLock()
+    {
+        _runningQueryLock.EnterReadLock();
+
+        return new DisposableAction(() => _runningQueryLock.ExitReadLock());
     }
 }
